intsets: move bitmap into a separate file
Release note: None
