+++ pkg/util/intsets/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "intsets",
     srcs = [
+        "bitmap.go",
         "fast_large.go",
         "fast_small.go",
         "fast_str.go",
